LPD-22028 Split the performance test class to export and import
