--- modules/apps/headless/headless-batch-engine/headless-batch-engine-test/src/testIntegration/java/com/liferay/headless/batch/engine/resource/v1_0/test/BatchExportImportPerformanceTest.java
@@ -1,407 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.headless.batch.engine.resource.v1_0.test;
-
-import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.headless.batch.engine.client.dto.v1_0.ExportTask;
-import com.liferay.headless.batch.engine.client.dto.v1_0.ImportTask;
-import com.liferay.headless.batch.engine.client.http.HttpInvoker;
-import com.liferay.headless.batch.engine.client.serdes.v1_0.ExportTaskSerDes;
-import com.liferay.headless.batch.engine.client.serdes.v1_0.ImportTaskSerDes;
-import com.liferay.petra.io.unsync.UnsyncByteArrayInputStream;
-import com.liferay.petra.string.StringBundler;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.json.JSONFactoryUtil;
-import com.liferay.portal.kernel.json.JSONUtil;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.servlet.HttpHeaders;
-import com.liferay.portal.kernel.test.randomizerbumpers.UniqueStringRandomizerBumper;
-import com.liferay.portal.kernel.test.rule.AggregateTestRule;
-import com.liferay.portal.kernel.test.rule.AssumeTestRule;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
-import com.liferay.portal.kernel.util.ContentTypes;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.LinkedHashMapBuilder;
-import com.liferay.portal.kernel.util.PropertiesUtil;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-
-import java.io.Closeable;
-import java.io.IOException;
-import java.io.InputStream;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Properties;
-import java.util.zip.ZipInputStream;
-
-import org.junit.Assert;
-import org.junit.Assume;
-import org.junit.BeforeClass;
-import org.junit.ClassRule;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-/**
- * @author Vendel Toreki
- */
-@RunWith(Arquillian.class)
-public class BatchExportImportPerformanceTest {
-
-	@ClassRule
-	@Rule
-	public static final AggregateTestRule aggregateTestRule =
-		new AggregateTestRule(
-			new AssumeTestRule("assume"), new LiferayIntegrationTestRule());
-
-	public static void assume() {
-		Assume.assumeTrue(Validator.isNull(System.getenv("JENKINS_HOME")));
-	}
-
-	@BeforeClass
-	public static void setUpClass() throws Exception {
-		Class<?> clazz = BatchExportImportPerformanceTest.class;
-
-		Properties properties = PropertiesUtil.load(
-			clazz.getResourceAsStream(
-				"dependencies/batch-export-import-performance.properties"),
-			"UTF-8");
-
-		_recordsCount = GetterUtil.getInteger(
-			properties.getProperty("records.count"));
-
-		_jsonTemplates = LinkedHashMapBuilder.put(
-			"com.liferay.headless.admin.user.dto.v1_0.UserAccount",
-			_createUserAccountJSONTemplate()
-		).build();
-	}
-
-	@Test
-	public void testImportAndExportTask() throws Exception {
-		for (String className : _jsonTemplates.keySet()) {
-			if (_log.isInfoEnabled()) {
-				_log.info("ClassName: " + className);
-			}
-
-			_testPostImportTask(className);
-
-			_testPostExportTask(className);
-		}
-	}
-
-	private static String _createUserAccountJSONTemplate() throws Exception {
-		return JSONUtil.put(
-			"additionalName", ""
-		).put(
-			"alternateName", "[$ALTERNATE_NAME$]"
-		).put(
-			"birthDate", "1977-01-01T00:00:00Z"
-		).put(
-			"customFields", JSONFactoryUtil.createJSONArray()
-		).put(
-			"dashboardURL", ""
-		).put(
-			"dateCreated", "2021-05-19T16:04:46Z"
-		).put(
-			"dateModified", "2021-05-19T16:04:46Z"
-		).put(
-			"emailAddress", "[$EMAIL_ADDRESS$]"
-		).put(
-			"familyName", "[$FAMILY_NAME$]"
-		).put(
-			"givenName", "[$GIVEN_NAME$]"
-		).put(
-			"jobTitle", ""
-		).put(
-			"keywords", JSONFactoryUtil.createJSONArray()
-		).put(
-			"name", "[$GIVEN_NAME$] [$FAMILY_NAME$]"
-		).put(
-			"organizationBriefs", JSONFactoryUtil.createJSONArray()
-		).put(
-			"profileURL", ""
-		).put(
-			"roleBriefs",
-			JSONUtil.put(
-				JSONUtil.put(
-					"id", 20113
-				).put(
-					"name", "User"
-				))
-		).put(
-			"siteBriefs",
-			JSONUtil.put(
-				JSONUtil.merge(
-					JSONUtil.put(
-						"id", 20127
-					).put(
-						"name", "Global"
-					),
-					JSONUtil.put(
-						"id", 20125
-					).put(
-						"name", "Guest"
-					)))
-		).put(
-			"userAccountContactInformation",
-			JSONUtil.put(
-				"emailAddresses", JSONFactoryUtil.createJSONArray()
-			).put(
-				"facebook", ""
-			).put(
-				"postalAddresses", JSONFactoryUtil.createJSONArray()
-			).put(
-				"skype", ""
-			).put(
-				"sms", ""
-			).put(
-				"telephones", JSONFactoryUtil.createJSONArray()
-			).put(
-				"twitter", ""
-			).put(
-				"webUrls", JSONFactoryUtil.createJSONArray()
-			)
-		).toString();
-	}
-
-	private String _createBatchJSON(String className, int recordsCount) {
-		StringBundler batchJsonSB = new StringBundler();
-
-		batchJsonSB.append(StringPool.OPEN_BRACKET);
-
-		for (int i = 0; i < recordsCount; i++) {
-			String alternateName = RandomTestUtil.randomString(
-				8, UniqueStringRandomizerBumper.INSTANCE);
-
-			String json = StringUtil.replace(
-				_jsonTemplates.get(className), "[$ALTERNATE_NAME$]",
-				alternateName);
-
-			json = StringUtil.replace(
-				json, "[$FAMILY_NAME$]",
-				StringUtil.getTitleCase(
-					RandomTestUtil.randomString(
-						8, UniqueStringRandomizerBumper.INSTANCE),
-					true, ""));
-
-			json = StringUtil.replace(
-				json, "[$GIVEN_NAME$]",
-				StringUtil.getTitleCase(
-					RandomTestUtil.randomString(
-						8, UniqueStringRandomizerBumper.INSTANCE),
-					true, ""));
-
-			batchJsonSB.append(
-				StringUtil.replace(
-					json, "[$EMAIL_ADDRESS$]",
-					StringBundler.concat(
-						alternateName, "@", RandomTestUtil.randomString(),
-						".com")));
-
-			if (i < (recordsCount - 1)) {
-				batchJsonSB.append(StringPool.COMMA);
-			}
-		}
-
-		batchJsonSB.append(StringPool.CLOSE_BRACKET);
-
-		return batchJsonSB.toString();
-	}
-
-	private String _getHttpResponseContent(String url) throws IOException {
-		HttpInvoker httpInvoker = HttpInvoker.newHttpInvoker();
-
-		httpInvoker.userNameAndPassword("test@liferay.com:test");
-		httpInvoker.httpMethod(HttpInvoker.HttpMethod.GET);
-		httpInvoker.path(url);
-
-		HttpInvoker.HttpResponse response = httpInvoker.invoke();
-
-		Assert.assertEquals(200, response.getStatusCode());
-
-		return response.getContent();
-	}
-
-	private Map<String, String> _splitClassName(String className) {
-		Map<String, String> classNamePartsMap = new HashMap<>();
-
-		if (className.contains("#")) {
-			String[] classNameParts = className.split("#");
-
-			classNamePartsMap.put("className", classNameParts[0]);
-			classNamePartsMap.put("taskItemDelegateName", classNameParts[1]);
-		}
-		else {
-			classNamePartsMap.put("className", className);
-		}
-
-		return classNamePartsMap;
-	}
-
-	private Closeable _startTimer() {
-		Thread thread = Thread.currentThread();
-
-		StackTraceElement stackTraceElement = thread.getStackTrace()[2];
-
-		String invokerName = StringBundler.concat(
-			stackTraceElement.getClassName(), StringPool.POUND,
-			stackTraceElement.getMethodName());
-
-		long startTime = System.currentTimeMillis();
-
-		return () -> {
-			if (_log.isInfoEnabled()) {
-				long totalTimeMillis = System.currentTimeMillis() - startTime;
-
-				double speed =
-					(double)(_recordsCount * 1000) / (double)totalTimeMillis;
-
-				_log.info(
-					StringBundler.concat(
-						invokerName, " used ", totalTimeMillis, " ms, for ",
-						_recordsCount, " records, speed: ",
-						String.format("%.3f", speed), " records/s"));
-			}
-		};
-	}
-
-	private void _testPostExportTask(String className) throws Exception {
-		Map<String, String> classNamePartsMap = _splitClassName(className);
-
-		try (Closeable closeable = _startTimer()) {
-			HttpInvoker httpInvoker = HttpInvoker.newHttpInvoker();
-
-			httpInvoker.header(
-				HttpHeaders.ACCEPT, ContentTypes.APPLICATION_JSON);
-			httpInvoker.userNameAndPassword("test@liferay.com:test");
-			httpInvoker.httpMethod(HttpInvoker.HttpMethod.POST);
-
-			StringBundler sb = new StringBundler(
-				classNamePartsMap.containsKey("taskItemDelegateName") ? 6 : 4);
-
-			sb.append("http://localhost:8080/o/headless-batch-engine/v1.0");
-			sb.append("/export-task/");
-			sb.append(classNamePartsMap.get("className"));
-			sb.append("/JSON");
-
-			if (classNamePartsMap.containsKey("taskItemDelegateName")) {
-				sb.append("?taskItemDelegateName=");
-				sb.append(classNamePartsMap.get("taskItemDelegateName"));
-			}
-
-			httpInvoker.path(sb.toString());
-
-			HttpInvoker.HttpResponse response = httpInvoker.invoke();
-
-			ExportTask exportTask = ExportTaskSerDes.toDTO(
-				response.getContent());
-
-			String externalReferenceCode =
-				exportTask.getExternalReferenceCode();
-
-			while (true) {
-				exportTask = ExportTaskSerDes.toDTO(
-					_getHttpResponseContent(
-						"http://localhost:8080/o/headless-batch-engine/v1.0" +
-							"/export-task/by-external-reference-code/" +
-								externalReferenceCode));
-
-				if (Objects.equals(
-						exportTask.getExecuteStatusAsString(), "COMPLETED")) {
-
-					break;
-				}
-				else if (Objects.equals(
-							exportTask.getExecuteStatusAsString(), "FAILED")) {
-
-					throw new AssertionError(exportTask.getErrorMessage());
-				}
-			}
-
-			httpInvoker = HttpInvoker.newHttpInvoker();
-
-			httpInvoker.header(
-				HttpHeaders.ACCEPT, ContentTypes.APPLICATION_OCTET_STREAM);
-			httpInvoker.userNameAndPassword("test@liferay.com:test");
-			httpInvoker.httpMethod(HttpInvoker.HttpMethod.GET);
-
-			httpInvoker.path(
-				StringBundler.concat(
-					"http://localhost:8080/o/headless-batch-engine/v1.0",
-					"/export-task/by-external-reference-code/",
-					externalReferenceCode, "/content"));
-
-			HttpInvoker.HttpResponse httpResponse = httpInvoker.invoke();
-
-			try (InputStream inputStream = new UnsyncByteArrayInputStream(
-					httpResponse.getBinaryContent())) {
-
-				ZipInputStream zipInputStream = new ZipInputStream(inputStream);
-
-				zipInputStream.getNextEntry();
-
-				StringUtil.read(zipInputStream);
-			}
-		}
-	}
-
-	private void _testPostImportTask(String className) throws Exception {
-		Map<String, String> classNamePartsMap = _splitClassName(className);
-
-		String json = _createBatchJSON(className, _recordsCount);
-
-		try (Closeable closeable = _startTimer()) {
-			HttpInvoker httpInvoker = HttpInvoker.newHttpInvoker();
-
-			httpInvoker.body(json, "application/json");
-			httpInvoker.userNameAndPassword("test@liferay.com:test");
-			httpInvoker.httpMethod(HttpInvoker.HttpMethod.POST);
-			httpInvoker.path(
-				StringBundler.concat(
-					"http://localhost:8080/o/headless-batch-engine/v1.0",
-					"/import-task/", classNamePartsMap.get("className"),
-					"?createStrategy=INSERT"));
-
-			HttpInvoker.HttpResponse response = httpInvoker.invoke();
-
-			ImportTask importTask = ImportTaskSerDes.toDTO(
-				response.getContent());
-
-			String externalReferenceCode =
-				importTask.getExternalReferenceCode();
-
-			while (true) {
-				importTask = ImportTaskSerDes.toDTO(
-					_getHttpResponseContent(
-						"http://localhost:8080/o/headless-batch-engine/v1.0" +
-							"/import-task/by-external-reference-code/" +
-								externalReferenceCode));
-
-				if (Objects.equals(
-						importTask.getExecuteStatusAsString(), "COMPLETED")) {
-
-					break;
-				}
-				else if (Objects.equals(
-							importTask.getExecuteStatusAsString(), "FAILED")) {
-
-					throw new AssertionError(importTask.getErrorMessage());
-				}
-			}
-		}
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		BatchExportImportPerformanceTest.class);
-
-	private static Map<String, String> _jsonTemplates;
-	private static int _recordsCount;
-
-}
