+++ modules/apps/headless/headless-batch-engine/headless-batch-engine-test/src/testIntegration/resources/META-INF/module-log4j.xml
@@ -2,7 +2,6 @@
 
 <Configuration strict="true">
 	<Loggers>
-		<Logger level="INFO" name="com.liferay.headless.batch.engine.resource.v1_0.test.ExportTaskResourceTest" />
-		<Logger level="INFO" name="com.liferay.headless.batch.engine.resource.v1_0.test.BatchExportImportPerformanceTest" />
+		<Logger level="INFO" name="com.liferay.headless.batch.engine.resource.v1_0.test" />
 	</Loggers>
 </Configuration>
